LogIn and SignIn fixes and remodeling
In this commit, changes to the way usuarios and almacenes are stored in the Firebase, changes to the object Models include adding Name to Usuarios and ID's to both usuarios and almacenes, this was added in order to make the future update of products easier. Remodeling of signin and login menus, consist of adding field for usuario name, resizing of components and adding buttons to facilitate app navigation.

CAMILO RODRIGUEZ GOMEZ